+++ Code/Scripts/UI/Shop/PlayerStatsDisplay.cs
@@ -56,7 +56,7 @@ public partial class PlayerStatsDisplay : Control
     /// Updates the health label.
     /// </summary>
     /// <param name="maxHealth">Is ignored, only needed beacuse of Event.</param>
-    public void UpdateHealthLabel(int maxHealth)
+    public void UpdateHealthLabel(float maxHealth)
     {
         _HPLabel.Text = $"{PlayerManager.Instance.PlayerHealthComponent.currentHP}/{PlayerManager.Instance.PlayerHealthComponent.maxHP}";
     }
